requirements.txt
Flask
passlib
google-generativeai
bcrypt